Cleanup step (sweep/tick) directory structure
