+++ packages/clock/src/math.tsx
@@ -1,7 +1,3 @@
-import { DateTime } from 'luxon'
-
-import type { StepStyle, HMS } from './types'
-
 export function hourToDegree(hour: number): number {
   return hour * 30
 }
@@ -24,35 +20,4 @@ export function degreeToCos(degree: number): number {
 
 export function degreeToRadian(degree: number): number {
   return degree * (Math.PI / 180)
-}
-
-export function calcHMS(
-  date: Date,
-  timezone: string,
-  stepStyle: StepStyle
-): HMS {
-  switch (stepStyle) {
-    case 'tick':
-      return tickHMS(date, timezone)
-    case 'sweep':
-      return sweepHMS(date, timezone)
-  }
-}
-
-export function tickHMS(date: Date, timezone: string): HMS {
-  const datetime = DateTime.fromJSDate(date)
-  const dt = datetime.setZone(timezone)
-  const hour = dt.hour
-  const minute = dt.minute
-  const second = dt.second
-  return { hour, minute, second }
-}
-
-export function sweepHMS(date: Date, timezone: string): HMS {
-  const datetime = DateTime.fromJSDate(date)
-  const dt = datetime.setZone(timezone)
-  const hour = dt.hour + dt.minute / 60 + dt.second / 3600
-  const minute = dt.minute + dt.second / 60
-  const second = dt.second + dt.millisecond / 1000
-  return { hour, minute, second }
 }
